Move NftRewrads to shared components dir
