+++ src/components/NftRewards/NftPreview.tsx
