+++ src/components/NftRewards/NftRewardsSection.tsx
