+++ src/components/NftRewards/RewardTier.tsx
